+++ components/src/main/java/org/patternfly/component/menu/PopperMultiMenuToggleMenu.java
@@ -35,7 +35,7 @@ import static org.patternfly.component.SelectionMode.multi;
  *
  * @param <B> the type of the builder for this component
  */
-@Deprecated(since = "0.7.0", forRemoval = true)
+@Deprecated
 abstract class PopperMultiMenuToggleMenu<B extends TypedBuilder<HTMLElement, B>> extends PopperMenuToggleMenu<B> {
 
     // ------------------------------------------------------ instance