+++ components/src/main/java/org/patternfly/component/menu/PopperSingleSelect.java
@@ -23,7 +23,7 @@ import org.patternfly.popper.TriggerAction;
  *
  * @deprecated This implementation uses the Popper.js API, which is due to be removed soon. Use {@link SingleSelect} instead.
  */
-@Deprecated(since = "0.7.0", forRemoval = true)
+@Deprecated
 public class PopperSingleSelect extends PopperSingleMenuToggleMenu<PopperSingleSelect> {
 
     // ------------------------------------------------------ factory