+++ core/src/main/java/org/patternfly/popper/PopperJs.java
@@ -27,7 +27,7 @@ import static jsinterop.annotations.JsPackage.GLOBAL;
  *
  * @see <a href="https://popper.js.org/docs/v2/">https://popper.js.org/docs/v2/</a>
  */
-@Deprecated(since = "0.7.0", forRemoval = true)
+@Deprecated
 @JsType(isNative = true, name = "Popper", namespace = GLOBAL)
 class PopperJs {